--- src/main/php/PDepend/Source/Language/PHP/PHPParserVersion55.php
@@ -1,337 +1,0 @@
-<?php
-/**
- * This file is part of PDepend.
- *
- * PHP Version 5
- *
- * Copyright (c) 2008-2017 Manuel Pichler <mapi@pdepend.org>.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- *   * Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *
- *   * Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in
- *     the documentation and/or other materials provided with the
- *     distribution.
- *
- *   * Neither the name of Manuel Pichler nor the names of his
- *     contributors may be used to endorse or promote products derived
- *     from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- * @copyright 2008-2017 Manuel Pichler. All rights reserved.
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- *
- * @since 2.3
- */
-
-namespace PDepend\Source\Language\PHP;
-
-use PDepend\Source\AST\ASTArray;
-use PDepend\Source\AST\ASTClassFqnPostfix;
-use PDepend\Source\AST\ASTLiteral;
-use PDepend\Source\AST\ASTNode;
-use PDepend\Source\Parser\UnexpectedTokenException;
-use PDepend\Source\Tokenizer\Tokens;
-
-/**
- * Concrete parser implementation that supports features up to PHP version 5.5.
- *
- * @copyright 2008-2017 Manuel Pichler. All rights reserved.
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- *
- * @since 2.3
- */
-abstract class PHPParserVersion55 extends AbstractPHPParser
-{
-    /**
-     * Parses a full qualified class name postfix.
-     *
-     * parseFullQualifiedClassNamePostfix() exists since 2.0.0 and have been customized for PHP 5.5 since 2.6.0.
-     *
-     * @return ASTClassFqnPostfix
-     *
-     * @since 2.0.0
-     */
-    protected function parseFullQualifiedClassNamePostfix()
-    {
-        $this->tokenStack->push();
-
-        $this->consumeToken(Tokens::T_CLASS_FQN);
-
-        return $this->setNodePositionsAndReturn(
-            $this->builder->buildAstClassFqnPostfix()
-        );
-    }
-
-    /**
-     * Will return <b>true</b> if the given <b>$tokenType</b> is a valid class
-     * name part.
-     *
-     * @param int $tokenType The type of a parsed token.
-     *
-     * @return bool
-     *
-     * @since  0.10.6
-     */
-    protected function isClassName($tokenType)
-    {
-        switch ($tokenType) {
-            case Tokens::T_NULL:
-            case Tokens::T_TRUE:
-            case Tokens::T_FALSE:
-            case Tokens::T_STRING:
-            case Tokens::T_READONLY:
-                return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * @param int $tokenType
-     *
-     * @return bool
-     */
-    protected function isConstantName($tokenType)
-    {
-        return $this->isFunctionName($tokenType);
-    }
-
-    /**
-     * @param int $tokenType
-     *
-     * @return bool
-     */
-    protected function isMethodName($tokenType)
-    {
-        return $this->isFunctionName($tokenType);
-    }
-
-    /**
-     * Tests if the give token is a valid function name in the supported PHP
-     * version.
-     *
-     * @param int $tokenType
-     *
-     * @return bool
-     *
-     * @since 2.3
-     */
-    protected function isFunctionName($tokenType)
-    {
-        switch ($tokenType) {
-            case Tokens::T_STRING:
-            case Tokens::T_NULL:
-            case Tokens::T_SELF:
-            case Tokens::T_TRUE:
-            case Tokens::T_FALSE:
-            case Tokens::T_PARENT:
-                return true;
-        }
-        return false;
-    }
-
-    /**
-     * Tests if the given token type is a reserved keyword in the supported PHP
-     * version.
-     *
-     * @param int $tokenType
-     *
-     * @return bool
-     */
-    protected function isKeyword($tokenType)
-    {
-        switch ($tokenType) {
-            case Tokens::T_CLASS:
-            case Tokens::T_TRAIT:
-            case Tokens::T_CALLABLE:
-            case Tokens::T_INSTEADOF:
-            case Tokens::T_INTERFACE:
-                return true;
-        }
-        return false;
-    }
-
-    /**
-     * Tests if the given token type is a valid type hint in the supported
-     * PHP version.
-     *
-     * @param int $tokenType
-     *
-     * @return bool
-     *
-     * @since 1.0.0
-     */
-    protected function isTypeHint($tokenType)
-    {
-        switch ($tokenType) {
-            case Tokens::T_CALLABLE:
-                return true;
-            default:
-                return parent::isTypeHint($tokenType);
-        }
-    }
-    protected function parseTypeHint()
-    {
-        switch ($this->tokenizer->peek()) {
-            case Tokens::T_CALLABLE:
-                $this->consumeToken(Tokens::T_CALLABLE);
-                $type = $this->builder->buildAstTypeCallable();
-                break;
-            default:
-                $type = parent::parseTypeHint();
-                break;
-        }
-        return $type;
-    }
-
-    /**
-     * Tests if the next token is a valid array start delimiter in the supported
-     * PHP version.
-     *
-     * @return bool
-     *
-     * @since 1.0.0
-     */
-    protected function isArrayStartDelimiter()
-    {
-        switch ($this->tokenizer->peek()) {
-            case Tokens::T_ARRAY:
-            case Tokens::T_SQUARED_BRACKET_OPEN:
-                return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Parses a php array declaration.
-     *
-     * @param bool $static
-     *
-     * @return ASTArray
-     *
-     * @since 1.0.0
-     */
-    protected function parseArray(ASTArray $array, $static = false)
-    {
-        switch ($this->tokenizer->peek()) {
-            case Tokens::T_SQUARED_BRACKET_OPEN:
-                $this->consumeToken(Tokens::T_SQUARED_BRACKET_OPEN);
-                $this->parseArrayElements($array, Tokens::T_SQUARED_BRACKET_CLOSE, $static);
-                $this->consumeToken(Tokens::T_SQUARED_BRACKET_CLOSE);
-                break;
-            case Tokens::T_ARRAY:
-                $this->consumeToken(Tokens::T_ARRAY);
-                $this->consumeComments();
-                $this->consumeToken(Tokens::T_PARENTHESIS_OPEN);
-                $this->parseArrayElements($array, Tokens::T_PARENTHESIS_CLOSE, $static);
-                $this->consumeToken(Tokens::T_PARENTHESIS_CLOSE);
-                break;
-            default:
-                throw $this->getUnexpectedNextTokenException();
-        }
-
-        return $array;
-    }
-
-    /**
-     * Parses an integer value.
-     *
-     * @throws UnexpectedTokenException
-     *
-     * @return ASTLiteral
-     */
-    protected function parseIntegerNumber()
-    {
-        $token = $this->consumeToken(Tokens::T_LNUMBER);
-        $number = $token->image;
-
-        while ($next = $this->addTokenToStackIfType(Tokens::T_STRING)) {
-            $number .= $next->image;
-        }
-
-        if ('0' !== substr($number, 0, 1)) {
-            goto BUILD_LITERAL;
-        }
-
-        if (Tokens::T_STRING !== $this->tokenizer->peek()) {
-            goto BUILD_LITERAL;
-        }
-
-        $token1 = $this->consumeToken(Tokens::T_STRING);
-        if (0 === preg_match('(^b[01]+$)i', $token1->image)) {
-            throw new UnexpectedTokenException(
-                $token1,
-                $this->tokenizer->getSourceFile()
-            );
-        }
-
-        $number .= $token1->image;
-        $token->endLine = $token1->endLine;
-        $token->endColumn = $token1->endColumn;
-
-        BUILD_LITERAL:
-
-        $literal = $this->builder->buildAstLiteral($number);
-        $literal->configureLinesAndColumns(
-            $token->startLine,
-            $token->endLine,
-            $token->startColumn,
-            $token->endColumn
-        );
-
-        return $literal;
-    }
-
-    /**
-     * Parses the class expr syntax supported since PHP 5.4.
-     *
-     * @return ASTNode
-     *
-     * @since 2.3
-     */
-    protected function parsePostfixIdentifier()
-    {
-        switch ($this->tokenizer->peek()) {
-            case Tokens::T_CURLY_BRACE_OPEN:
-                $node = $this->parseCompoundExpression();
-                break;
-            default:
-                $node = parent::parsePostfixIdentifier();
-                break;
-        }
-        return $this->parseOptionalIndexExpression($node);
-    }
-
-    /**
-     * @return ASTNode
-     */
-    protected function parseOptionalExpressionForVersion()
-    {
-        switch ($this->tokenizer->peek()) {
-            case Tokens::T_TRAIT_C:
-                return $this->parseConstant();
-            default:
-                return parent::parseOptionalExpressionForVersion();
-        }
-    }
-}
